scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "google-generativeai>=0.8.5",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "supabase>=2.25.0",
]
